adding support for specifying multiple queues.
In some cases you can not use the prefix to check multiple queus with the same values (such as having a sufix to have it be identidified as a dead letter queue). In these scenerios you might still want to be able to give it multiple queues to reduce the number of times you must invoke as initializing aws-sdk is expensive.

relates to:
- #24
- #27