Feature/allow zero numerator (#266)
* Add ability for numerator to be empty

CloudWatch returns empty data if there is no metric for it. This allows
us to set a default (0) to replace for the numeratort.

Remove the --no-data-ok flag because it doesn't make sense with the
above option. Instead we chang eit to --no-denominator-data-ok. If there
is no denominator data we will either `unknown` or `ok`.

Also refactored some to reuse more code. May refactor more.

* Allow backwards compatibility and add some tests

* Move #composite_check out of library, and update tests

* Split out some logic for easier legibility

* Fix rubo by disabling and adding changelog