Add `check-efs-metric.rb` plugin
This plugin checks a CloudWatch metric for the named EFS filesystem.
An optional `-N` argument will return an `ok` result if the filesystem
is not found - useful for running the same check across environments
where filesystems may not always exist.

This plugin is based on `check-cloudwatch-metric.rb` and queries the
`AWS/EFS` CloudWatch namespace using the EFS filesystem ID determined
from its name.

Here is a sample run of the plugin:

```
$ check-efs-metric.rb --region us-west-2 --name app_nfs --metric PercentIOLimit --operator greater --warning 60.0 --critical 80.0
EFSMetric OK: PercentIOLimit for app_nfs is 0.0: comparison=greater, will alarm at 60.0
```