Run check against most recent s3 object in list of matching objects (#282)
* Added flag to allow checking against multiple objects

* check-rds-pending: Handle no clusters properly

* prep for 11.5.1 release

* Provide the --fetch-age option for the RDS metrics collector

This allows to get metrics in the past in a more automated way than setting
the end_time directly.

* prep for 11.6.0 release

* Up the period from 60 to 120

When `config[:period]` is at 60, `client.get_metric_statistics()` will return from 0 to 2 results, which can result in a false positive if the returned value for resp.datapoints is a `[]`.

Setting this value to 120 will consistently return 2-3 results. CloudWatch API does not guarantee that the values returned will be in time order, so I believe this should be a good compromise. `period` has to be in a multiple of 60, therefore 120 is the next best value.

* Updated changelog to reflect ebs burst check issue

* prep for 12.0.0 release

* Add `check-efs-metric.rb` plugin

This plugin checks a CloudWatch metric for the named EFS filesystem.
An optional `-N` argument will return an `ok` result if the filesystem
is not found - useful for running the same check across environments
where filesystems may not always exist.

This plugin is based on `check-cloudwatch-metric.rb` and queries the
`AWS/EFS` CloudWatch namespace using the EFS filesystem ID determined
from its name.

Here is a sample run of the plugin:

```
$ check-efs-metric.rb --region us-west-2 --name app_nfs --metric PercentIOLimit --operator greater --warning 60.0 --critical 80.0
EFSMetric OK: PercentIOLimit for app_nfs is 0.0: comparison=greater, will alarm at 60.0
```

* add changelog for #292

* prep for 12.1.0 release

* Add check for Direct Connect virtual interfaces

* prep for 12.2.0 release

* Added flag to allow checking against multiple objects

* Add pagination

* Change long option name